Fix import error (#96)
* move rename functions

* add json_utils file

* move tests

* move create_sample_run into test_utils

* comments

* apply new import to tma notebook

* pycode style

* trigger new build
