+++ toffy/rename_fovs_test.py
@@ -1,42 +1,10 @@
 import tempfile
-import json
 import os
 import pytest
 
 from ark.utils import io_utils
 from toffy import rename_fovs as rf
-
-
-def create_sample_run(name_list, run_order_list, scan_count_list, create_json=False, bad=False):
-    fov_list = []
-    sample_run = {"fovs": fov_list}
-
-    # set up dictionary
-    for name, run_order, scan_count in zip(name_list, run_order_list, scan_count_list):
-        ex_fov = {
-            "scanCount": scan_count,
-            "runOrder": run_order,
-            "name": name
-        }
-        fov_list.append(ex_fov)
-
-    # delete name key if one is not provided
-    for fov in sample_run.get('fovs', ()):
-        if fov.get('name') is None:
-            del fov['name']
-
-    # create bad dictionary
-    if bad:
-        sample_run['bad key'] = sample_run['fovs']
-        del sample_run['fovs']
-
-    # create json file for the data
-    if create_json:
-        temp = tempfile.NamedTemporaryFile(mode="w", delete=False)
-        json.dump(sample_run, temp)
-        return temp.name
-
-    return sample_run
+from toffy import test_utils
 
 
 def create_sample_fov_dirs(fovs, base_dir):
@@ -50,21 +18,6 @@ def remove_fov_dirs(base_dir):
     fovs = io_utils.list_folders(base_dir)
     for fov in fovs:
         os.rmdir(os.path.join(base_dir, fov))
-
-
-def test_rename_missing_fovs():
-    # data with missing names
-    ex_name = ['custom_1', None, 'custom_2', 'custom_3', None]
-    ex_run_order = list(range(1, 6))
-    ex_scan_count = list(range(1, 6))
-
-    # create a dict with the sample data
-    ex_run = create_sample_run(ex_name, ex_run_order, ex_scan_count)
-
-    # test that missing names are given a placeholder
-    ex_run = rf.rename_missing_fovs(ex_run)
-    for fov in ex_run.get('fovs', ()):
-        assert fov.get('name') is not None
 
 
 def test_rename_fov_dirs():
@@ -90,10 +43,11 @@ def test_rename_fov_dirs():
         ex_scan_count = [1, 2, 1]
 
         # create a json path with the sample data
-        ex_run_path = create_sample_run(ex_name, ex_run_order, ex_scan_count, True)
+        ex_run_path = test_utils.create_sample_run(ex_name, ex_run_order, ex_scan_count, True)
 
         # bad sample run data
-        bad_run = create_sample_run(ex_name, ex_run_order, ex_scan_count, True, bad=True)
+        bad_run = test_utils.create_sample_run(ex_name, ex_run_order, ex_scan_count,
+                                               True, bad=True)
 
         # bad run file data should raise an error
         with pytest.raises(KeyError):
