+++ toffy/tiling_utils.py
@@ -17,7 +17,7 @@ import warnings
 
 from dataclasses import dataclass
 
-from toffy import settings, rename_fovs
+from toffy import settings, json_utils
 from ark.utils import misc_utils
 
 
@@ -357,7 +357,7 @@ def set_tiled_region_params(region_corners_path):
     # read in the region corners data
     with open(region_corners_path, 'r', encoding='utf-8') as flf:
         tiled_region_corners = json.load(flf)
-    tiled_region_corners = rename_fovs.rename_missing_fovs(tiled_region_corners)
+    tiled_region_corners = json_utils.rename_missing_fovs(tiled_region_corners)
 
     # define the parameter dict to return
     tiling_params = {}
@@ -667,7 +667,7 @@ def generate_tma_fov_list(tma_corners_path, num_fov_row, num_fov_col):
     # read in tma_corners_path
     with open(tma_corners_path, 'r', encoding='utf-8') as flf:
         tma_corners = json.load(flf)
-    tma_corners = rename_fovs.rename_missing_fovs(tma_corners)
+    tma_corners = json_utils.rename_missing_fovs(tma_corners)
 
     # a TMA can only be defined by four FOVs, one for each corner
     if len(tma_corners['fovs']) != 4:
@@ -697,36 +697,6 @@ def generate_tma_fov_list(tma_corners_path, num_fov_row, num_fov_col):
         fov_regions[fov_names[index]] = (xi, yi)
 
     return fov_regions
-
-
-def rename_duplicate_fovs(tma_fovs):
-    """Identify and rename duplicate FOV names in `fov_list`
-
-    For a given FOV name, the subsequent duplicates get renamed `{FOV}_duplicate{n}`
-
-    Args:
-        tma_fovs (dict):
-            The TMA run JSON, should contain a `'fovs'` key defining the list of FOVs
-
-    Returns:
-        dict:
-            The same run JSON with the FOVs renamed to account for duplicates
-    """
-
-    # used for identifying the number of times each FOV was found
-    fov_count = {}
-
-    # iterate over each FOV
-    for fov in tma_fovs['fovs']:
-        if fov['name'] not in fov_count:
-            fov_count[fov['name']] = 0
-
-        fov_count[fov['name']] += 1
-
-        if fov_count[fov['name']] > 1:
-            fov['name'] = '%s_duplicate%d' % (fov['name'], fov_count[fov['name']] - 1)
-
-    return tma_fovs
 
 
 def convert_stage_to_optical(coord, stage_optical_coreg_params):
